+++ src/basis/compiler/mcPrimitives.c
@@ -1,8 +1,8 @@
 
-#include "../../include/cfrtil.h"
+#include "../../include/cfrtil32.h"
 
 void
-MCP_MoveArrayN_To_Reg ( int64 reg, byte * array, int64 n )
+MCP_MoveArrayN_To_Reg ( int32 reg, byte * array, int32 n )
 {
     // pop reg, array, n
     // move array to reg
